feat(global): JWT 신뢰 경계 개선 - Gateway 헤더 주입 전환 - hoppingmall-common: GatewayHeaderAuthenticationFilter 추가 (x-user-id/x-user-role 헤더 → SecurityContext) - 6개 서비스: JwtAuthenticationFilter + JwtTokenParser 제거 (Base64 디코딩 보안 취약점 해소) - 6개 서비스 SecurityConfig: GatewayHeaderAuthenticationFilter로 교체 - Gateway에서 JWT 서명 검증 완료 후 내부 헤더만 신뢰하는 구조로 전환
