--- notification-service/src/main/kotlin/com/hoppingmall/notification/jwt/JwtAuthenticationFilter.kt
@@ -1,47 +1,0 @@
-package com.hoppingmall.notification.jwt
-
-import com.hoppingmall.common.UserPrincipal
-import jakarta.servlet.FilterChain
-import jakarta.servlet.http.HttpServletRequest
-import jakarta.servlet.http.HttpServletResponse
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
-import org.springframework.security.core.context.SecurityContextHolder
-import org.springframework.stereotype.Component
-import org.springframework.web.filter.OncePerRequestFilter
-
-@Component
-class JwtAuthenticationFilter(
-    private val jwtTokenParser: JwtTokenParser
-) : OncePerRequestFilter() {
-
-    override fun doFilterInternal(
-        request: HttpServletRequest,
-        response: HttpServletResponse,
-        filterChain: FilterChain
-    ) {
-        val token = resolveToken(request)
-
-        if (token != null) {
-            val userId = jwtTokenParser.parseUserId(token)
-            val role = jwtTokenParser.parseRole(token) ?: "BUYER"
-
-            if (userId != null) {
-                val userPrincipal = UserPrincipal.of(userId, role)
-                val authentication = UsernamePasswordAuthenticationToken(
-                    userPrincipal, null, userPrincipal.authorities
-                )
-                SecurityContextHolder.getContext().authentication = authentication
-            }
-        }
-
-        filterChain.doFilter(request, response)
-    }
-
-    private fun resolveToken(request: HttpServletRequest): String? {
-        val bearerToken = request.getHeader("Authorization")
-        if (bearerToken != null && bearerToken.startsWith("Bearer ")) {
-            return bearerToken.substring(7)
-        }
-        return null
-    }
-}
