--- user-service/src/main/kotlin/com/hoppingmall/user/auth/JwtAuthenticationFilter.kt
@@ -1,47 +1,0 @@
-package com.hoppingmall.user.auth
-
-import com.hoppingmall.user.auth.domain.repository.AccessTokenBlacklistRepository
-import jakarta.servlet.FilterChain
-import jakarta.servlet.http.HttpServletRequest
-import jakarta.servlet.http.HttpServletResponse
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken
-import org.springframework.security.core.context.SecurityContextHolder
-import org.springframework.security.web.authentication.WebAuthenticationDetailsSource
-import org.springframework.stereotype.Component
-import org.springframework.web.filter.OncePerRequestFilter
-
-@Component
-class JwtAuthenticationFilter(
-    private val tokenProvider: TokenProvider,
-    private val accessTokenBlacklistRepository: AccessTokenBlacklistRepository
-) : OncePerRequestFilter() {
-
-    override fun doFilterInternal(
-        request: HttpServletRequest,
-        response: HttpServletResponse,
-        filterChain: FilterChain
-    ) {
-        val token = resolveToken(request)
-
-        if (token != null && tokenProvider.validateToken(token) && !accessTokenBlacklistRepository.exists(token)) {
-            val userPrincipal = tokenProvider.getUserPrincipal(token)
-
-            val authentication = UsernamePasswordAuthenticationToken(
-                userPrincipal,
-                null,
-                userPrincipal.authorities
-            ).apply {
-                details = WebAuthenticationDetailsSource().buildDetails(request)
-            }
-
-            SecurityContextHolder.getContext().authentication = authentication
-        }
-
-        filterChain.doFilter(request, response)
-    }
-
-    private fun resolveToken(request: HttpServletRequest): String? {
-        val bearerToken = request.getHeader("Authorization") ?: return null
-        return if (bearerToken.startsWith("Bearer ")) bearerToken.substring(7) else null
-    }
-}
