+++ user-service/src/main/kotlin/com/hoppingmall/user/config/SecurityConfig.kt
@@ -1,7 +1,7 @@
 package com.hoppingmall.user.config
 
 import com.hoppingmall.common.config.InternalTokenFilter
-import com.hoppingmall.user.auth.JwtAuthenticationFilter
+import com.hoppingmall.common.config.GatewayHeaderAuthenticationFilter
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
 import org.springframework.security.config.annotation.web.builders.HttpSecurity
@@ -15,7 +15,7 @@ import org.springframework.security.web.authentication.UsernamePasswordAuthentic
 @Configuration
 @EnableWebSecurity
 class SecurityConfig(
-    private val jwtAuthenticationFilter: JwtAuthenticationFilter,
+    private val gatewayHeaderAuthenticationFilter: GatewayHeaderAuthenticationFilter,
     private val internalTokenFilter: InternalTokenFilter
 ) {
     @Bean
@@ -34,7 +34,7 @@ class SecurityConfig(
                     .anyRequest().authenticated()
             }
             .addFilterBefore(internalTokenFilter, UsernamePasswordAuthenticationFilter::class.java)
-            .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter::class.java)
+            .addFilterBefore(gatewayHeaderAuthenticationFilter, UsernamePasswordAuthenticationFilter::class.java)
         return http.build()
     }
